plugins/desktopchangeosd: fix per-output desktops
It always showed up on active screen, which is not always the best
behavior with per-output virtual desktops enabled. Consider this
scenario:

- Per-output desktops are enabled.
- There is a window on screen A which is not on it's current desktop.
- There is a task manager on screen B which includes tasks from all
  screens and desktops.
- The window is activated from the task manager on screen B, which
  causes screen A's desktop to change. However, B is the active screen,
  so the OSD previously showed up on screen B. Now it shows up on A.

There is still only one dialog shown at a time, even if desktops change
on multiple screens (e.g. per-output desktops are disabled). This is
handled by preferring the active screen.